requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.3
django-cors-headers==3.13.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
phonenumbers==8.13.6
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.6
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
